agentic_ai/autogen/requirements.txt
Flask
Faker
python-dotenv
openai
fastapi
fastmcp
flasgger
--prefer-binary
pyarrow
streamlit
uvicorn
autogen-agentchat==0.5.6
autogen-ext[azure,mcp,openai]==0.5.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-openai
ipykernel
pymongo
asyncpg
json_schema_to_pydantic